TASK #19076: Handle Login Button
